+++ launcher/sources/main.cpp
@@ -34,10 +34,12 @@ int runCLI(int argc, char* argv[]) {
   Solarus::Debug::set_abort_on_die(true);
   Solarus::Debug::set_show_popup_on_die(false);
 
-  // Run the Solarus main loop.
-  const Solarus::Arguments args(argc, argv);
-  Solarus::MainLoop(args).run();
+  Solarus::Arguments args;
+  for (int i = 0; i < argc; ++i) {
+    args.add_argument(QString::fromLocal8Bit(argv[i]).toUtf8().constData());
+  }
 
+  Solarus::MainLoop(args).run();
   return EXIT_SUCCESS;
 }