+++ tests/testing_quest/data/maps/bugs/1672_wrong_camera_position_on_started.lua
@@ -1,0 +1,13 @@
+local map = ...
+
+local sx, sy
+
+function map:on_started()
+  sx, sy = self:get_camera():get_position()
+end
+
+function map:on_opening_transition_finished()
+  local x, y = self:get_camera():get_position()
+  assert(x == sx and y == sy, "Camera position after the opening transition should be the same as position on start")
+  sol.main.exit()
+end